app/requirements.txt
openai==1.14.3
python-dotenv==1.0.0
langchain==0.0.346
tiktoken==0.5.1
chainlit==1.0.506
azure-data-tables==12.5.0
azure-ai-documentintelligence==1.0.0b3
# python-docx==1.1.2
bayoo-docx==0.2.20
# azure-ai-formrecognizer==3.3.3

